Lessons Learned deck
Lessons learned from the AdoptOpenJDK build farm.

This deck contains the basics of OpenJDK and AdoptOpenJDK, the
background and mission of AdoptOpenJDK, and a run-through of the
build pipeline, plus the tech used to put it together.

This deck is timed for 1.5 hours. A 2-hour slot is recommended, as
there were many questions and comments at both showings so far,
and it's very easy to run long with this deck.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>